+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\files\assertWritableDirectory
 - Adds oihana\core\files\createDirectory
 - Adds oihana\core\files\deleteDirectory
+- Adds oihana\core\files\getDirectory
 - Adds oihana\enums\FileExtension
 - Adds oihana\files\TarFile